Revert previous commit to int for default bool size.
intptr_t leads to verbose code during printf.


git-svn-id: http://svn.cri.ensmp.fr/svn/newgen/trunk@1281 80f2d23d-f6f1-0310-8ea7-b0e6515bd809

+++ src/genC/newgen_types.h
@@ -73,7 +73,7 @@ typedef uintptr_t _uint;
     #error newgen header not compatible with stdbool.h
 #endif
 
-typedef intptr_t bool; /* we cannot use an enum or stdbool because we need to be compatible with newgen */
+typedef int bool; /* we cannot use an enum or stdbool because we need to be compatible with newgen */
 
 #define false 0
 #define true 1
